fixing data model tests
